+++ src/syncData.ts
@@ -10,6 +10,7 @@ import type {
   ChangeStreamInsertDocument,
   Collection,
   Document,
+  ObjectId,
 } from 'mongodb'
 import { mapLeaves } from 'obj-walker'
 import type { QueueOptions } from 'prom-utils'
@@ -19,11 +20,12 @@ import type { Crate, ErrorResult, QueryResult } from './crate.js'
 import type {
   ConvertOptions,
   Events,
-  ImmutableOption,
+  OptimizationOptions,
   SyncOptions,
 } from './types.js'
 import {
   getFailedRecords,
+  partitionEvents,
   renameKeys,
   setDefaults,
   sumByRowcount,
@@ -70,7 +72,7 @@ export const initSync = (
   const handleResult = (
     result: QueryResult | ErrorResult,
     operationType: ChangeStreamDocument['operationType'],
-    doc: Document
+    _id: ObjectId
   ) => {
     debug('Change stream result %O', result)
     if ('rowcount' in result) {
@@ -80,7 +82,7 @@ export const initSync = (
         operationCounts: { [operationType]: 1 },
       })
     } else {
-      emit('error', { error: result, changeStream: true, failedDoc: doc })
+      emit('error', { error: result, changeStream: true, failedDoc: _id })
     }
   }
   /**
@@ -92,29 +94,31 @@ export const initSync = (
     try {
       if (doc.operationType === 'insert') {
         const document = mapper(doc.fullDocument)
+        const _id = doc.documentKey._id
         const result = await crate.insert(qualifiedName, document)
-        handleResult(result, doc.operationType, document)
+        handleResult(result, doc.operationType, _id)
       } else if (doc.operationType === 'update') {
         const document = doc.fullDocument ? mapper(doc.fullDocument) : {}
         const { updatedFields, removedFields } = doc.updateDescription
         const removed = removedFields && setDefaults(removedFields, null)
         const update = mapper({ ...updatedFields, ...removed })
         if (_.size(update)) {
+          const _id = doc.documentKey._id
           const result = await crate.upsert(qualifiedName, document, update)
-          handleResult(result, doc.operationType, document)
+          handleResult(result, doc.operationType, _id)
         }
       } else if (doc.operationType === 'replace') {
-        const id = doc.documentKey._id.toString()
+        const _id = doc.documentKey._id
         // Delete
-        await crate.deleteById(qualifiedName, id)
+        await crate.deleteById(qualifiedName, _id.toString())
         // Insert
         const document = mapper(doc.fullDocument)
         const result = await crate.insert(qualifiedName, document)
-        handleResult(result, doc.operationType, document)
+        handleResult(result, doc.operationType, _id)
       } else if (doc.operationType === 'delete') {
-        const id = doc.documentKey._id.toString()
-        const result = await crate.deleteById(qualifiedName, id)
-        handleResult(result, doc.operationType, { id })
+        const _id = doc.documentKey._id
+        const result = await crate.deleteById(qualifiedName, _id.toString())
+        handleResult(result, doc.operationType, _id)
       }
     } catch (e) {
       emit('error', { error: e, changeStream: true })
@@ -136,7 +140,7 @@ export const initSync = (
         const numInserted = sumByRowcount(1)(result.results)
         // -2 indicates failure
         const numFailed = sumByRowcount(-2)(result.results)
-        const failedDocs = getFailedRecords(result.results, documents)
+        const failedDocs = getFailedRecords(result.results, docs)
         emit('process', {
           success: numInserted,
           fail: numFailed,
@@ -154,20 +158,25 @@ export const initSync = (
   }
 
   const processChangeStream = (
-    options?: QueueOptions & ChangeStreamOptions & ImmutableOption
+    options?: QueueOptions & ChangeStreamOptions & OptimizationOptions
   ) =>
-    options?.immutable
-      ? // The collection is immutable and therefore only inserts will occur
-        sync.processChangeStream(
-          (docs) =>
-            processInsertRecords(
-              // Typscript hack
-              docs as unknown as ChangeStreamInsertDocument[],
-              'changeStream'
-            ),
-          // Ensure we only receive insert events
-          { ...options, operationTypes: ['insert'] }
-        )
+    options?.autoOptimizeInserts || options?.immutable
+      ? sync.processChangeStream(async (docs) => {
+          const partitions = partitionEvents(docs)
+          for (const partition of partitions) {
+            // We have more than one event so this is a grouped set of inserts
+            if (partition.length > 1) {
+              debug('Change stream insert batch of length %d', partition.length)
+              await processInsertRecords(
+                // We know these are going to be insert events
+                partition as unknown as ChangeStreamInsertDocument[],
+                'changeStream'
+              )
+            } else {
+              await processChangeStreamRecords(partition)
+            }
+          }
+        }, options)
       : sync.processChangeStream(processChangeStreamRecords, {
           ...options,
           // We can only handle one record at a time