+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.49.0
+
+-   **Deprecated** `immutable` option for `processChangeStream` in favor of `autoOptimizeInserts` option that works well in an insert-only
+    or insert-heavy scenario. If you only want `insert` events, you must pass `{operationTypes: ['insert']}`.
+-   Only return `_id`s for `failedDocs` and `failedDoc` in emitted events.
+-   Bumped dependencies.
+
 # 0.48.0
 
 -   **Breaking Change**: When there are multiple `overrides` that match the same